pyproject.toml
[project]
name = "constructive-airsim-ms"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "airsim>=1.8.1",
    "msgpack-rpc-python>=0.4.1",
    "tornado>=4.5,<6",            # msgpack-rpc-python incompatible with tornado>=6
    "openai>=1.30",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "grpcio>=1.64",
    "grpcio-tools>=1.64",
    "pymap3d>=3.1",
    "numpy>=1.26",
    "pydantic>=2.7",
    "structlog>=24.1",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/constructive_airsim_ms"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
